refactor: 중복 폴더 삭제 (lib, util, src/types)
