+++ src/features/auth/api/index.tsx
@@ -1,4 +1,4 @@
-import axiosInst from '@/util/axiosInst'
+import axiosInst from '@/lib/axiosInst'
 
 type SignUpResponseType = {
   response: {
