+++ src/features/reviews/hooks/useStoreReviewInfiniteQuery.test.tsx
@@ -1,5 +1,5 @@
 import { renderHook, waitFor } from '@testing-library/react'
-import { wrapper } from '@/util/test-utils'
+import { wrapper } from '@/lib/test-utils'
 import { useStoreReviewInfiniteQuery } from '.'
 
 describe('useStoreReviewInfiniteQuery', () => {
