--- src/types/store.ts
@@ -1,9 +1,0 @@
-export type Store = {
-  storeId: number
-  storeName: string
-  thumbnails: string[]
-  reviewRate: number
-  reviewCount: number
-  deliveryPrice: number
-  category: string
-}
